requirements.txt
absl-py==0.7.0
aiohttp==3.7.4.post0
alabaster==0.7.12
algopy==0.5.7
alpha-vantage==2.3.1
amqp==1.4.9
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.2
anyjson==0.3.3
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
astropy==3.1.2
async-timeout==3.0.1
asyncio==3.4.3
atomicwrites==1.3.0
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
billiard==3.3.0.23
bitarray==0.8.3
bitsets==0.8.2
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
blis==0.4.1
bokeh==1.0.4
bootstrapped==0.0.2
boto==2.49.0
Bottleneck==1.2.1
bs4==0.0.1
cachetools==4.2.2
catalogue==0.0.8
celery==3.1.25
certifi==2021.5.30
cffi==1.13.2
chardet==3.0.4
charset-normalizer==2.0.3
Click==7.0
cloudpickle==0.8.0
clyent==1.2.2
colorama==0.4.1
colorlover==0.2.1
comtypes==1.1.7
concepts==0.9.2
conda==4.7.12
conda-build==3.17.8
conda-package-handling==1.6.0
confuse==1.0.0
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.8
cs50==4.0.4
cssselect==1.1.0
cufflinks==0.14.6
cycler==0.10.0
cymem==2.0.2
Cython==0.29.5
cytoolz==0.9.0.1
dash==1.9.0
dash-core-components==1.8.0
dash-daq==0.3.3
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
dask==1.1.2
dataclasses==0.8
datashape==0.5.4
decorator==4.3.2
defusedxml==0.5.0
distributed==1.26.0
docutils==0.14
dtale==1.7.5
en-core-web-lg==2.2.5
en-core-web-md==2.2.5
en-core-web-sm==2.2.5
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
Faker==1.0.2
fastcache==1.0.2
features==0.5.12
fileconfig==0.5.10
filelock==3.0.10
findspark==1.4.2
Flask==1.1.1
Flask-Caching==1.8.0
Flask-Compress==1.4.0
Flask-Cors==3.0.7
Flask-Session==0.3.1
Flask-SSLify==0.1.5
future==0.17.1
gast==0.2.2
gevent==1.4.0
glob2==0.6
google-api-core==1.31.0
google-auth==1.33.1
google-cloud-appengine-logging==0.1.3
google-cloud-audit-log==0.1.0
google-cloud-bigquery==2.22.1
google-cloud-core==1.7.1
google-cloud-logging==2.5.0
google-cloud-pubsub==2.6.1
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
graphviz==0.17
greenlet==0.4.15
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
htmlmin==0.1.12
hyperlink==18.0.0
idna==3.2
idna-ssl==1.1.0
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.23
imutils==0.5.2
incremental==17.5.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.8
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
Jinja2==2.10.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.3
jupyterlab-launcher==0.13.1
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==18.0.0
kiwisolver==1.0.1
kombu==3.0.37
lazy-object-proxy==1.3.1
libarchive-c==2.8
libcst==0.3.19
llvmlite==0.27.0
locket==0.2.0
lxml==4.3.0
lz4==3.0.2
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
MDP==3.5
menuinst==1.4.14
missingno==0.4.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==5.1.0
multipledispatch==0.6.0
murmurhash==1.0.2
mypy-extensions==0.4.3
navigator-updater==0.2.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.4
numba==0.42.0
numdifftools==0.9.20
numexpr==2.6.8
numpy==1.15.4
numpydoc==0.8.0
odo==0.5.1
olefile==0.46
opencv-python==4.0.0.21
openpyxl==2.6.0
packaging==21.0
pandas==0.24.1
pandas-profiling==2.0.3
pandocfilters==1.4.2
parse==1.12.1
parso==0.3.4
partd==0.3.9
passlib==1.7.2
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
phik==0.9.8
pickleshare==0.7.5
Pillow==5.4.1
pkginfo==1.5.0.1
plac==0.9.6
playsound==1.2.2
plotly==3.4.2
pluggy==0.9.0
ply==3.11
preshed==3.0.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
proto-plus==1.19.0
protobuf==3.17.3
psutil==5.5.0
py==1.8.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyee==6.0.0
pyflakes==2.1.0
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.2.2
pymemcache==2.2.2
pyodbc==4.0.26
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
pyppeteer==0.0.25
pyquery==1.4.1
PySocks==1.6.8
pyspark==2.4.7
pytesseract==0.2.6
pytest==4.3.0
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.2.0
pytest-openfiles==0.3.2
pytest-pylint==0.14.0
pytest-remotedata==0.3.1
pytest-runner==4.4
python-dateutil==2.8.0
python-memcached==1.59
pytz==2021.1
PyWavelets==1.0.1
pywin32==223
pywinpty==0.5.5
PyYAML==5.4.1
pyzmq==17.1.2
QtAwesome==0.5.6
qtconsole==4.4.3
QtPy==1.6.0
redis==2.10.6
requests==2.26.0
requests-html==0.10.0
retrying==1.3.3
rope==0.12.0
rq==1.1.0
rsa==4.7.2
ruamel-yaml==0.15.46
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.16.0
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.7.1
spacy==2.2.3
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.11
sqlparse==0.3.0
srsly==0.2.0
statsmodels==0.9.0
sympy==1.3
tables==3.4.4
tblib==1.3.2
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.2
textblob==0.15.3
textblob-aptagger==0.2.0
thinc==7.3.1
toolz==0.9.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
Twisted==18.9.0
typed-ast==1.3.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
unicodecsv==0.14.1
urllib3==1.26.6
vine==1.3.0
virtualenv==16.7.6
w3lib==1.21.0
wasabi==0.4.2
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.0
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
win-unicode-console==0.5
win10toast==0.9
win10toast-persist==0.10.1
wincertstore==0.2
windows-curses==1.1
wrapt==1.11.1
xlrd==1.2.0
XlsxWriter==1.1.5
xlwings==0.15.3
xlwt==1.3.0
xmljson==0.2.0
xmltodict==0.12.0
yahoo-fin==0.8.2
yarl==1.6.3
yellowbrick==0.9.1
zict==0.1.3
zipp==0.6.0
zope.interface==4.6.0
